Rework all comments
