+++ rss_reader/setup.py
@@ -2,11 +2,11 @@ from setuptools import setup
 
 setup(
     name='rss_reader',
-    version='0.4',
+    version='1.0',
     description='Command-line RSS reader',
     author='Kolesnikov Viktor',
     url='https://github.com/MackDillan/PythonFinalTask',
-    install_requires=['beautifulsoup4', 'lxml', 'requests', 'peewee', 'xhtml2pdf', ],
+    install_requires=['beautifulsoup4', 'lxml', 'requests', 'peewee', 'xhtml2pdf', 'python-dateutil', ],
     python_requires='>=3.8',
     packages=['rss_reader'],
     entry_points={
