+++ rss_reader/tests/test_errors.py
@@ -1,6 +1,6 @@
 import unittest
 import io
-from rss_reader.rss_reader import main
+from rss_reader.reader import main
 from contextlib import redirect_stdout
 
 
@@ -21,14 +21,13 @@ class TestRssReaderErrors(unittest.TestCase):
         with self.assertRaises(SystemExit):
             with io.StringIO() as buf, redirect_stdout(buf):
                 main(args=['url.com'])
-                self.assertEqual(buf.getvalue(),
-                                 'Invalid URL url.com: No schema supplied. Perhaps you meant http:url.com?\n')
+                self.assertEqual(buf.getvalue(), 'Invalid URL {url}. Please correct the URL and try again\n')
 
     def test_invalid_url(self):
         with self.assertRaises(SystemExit):
             with io.StringIO() as buf, redirect_stdout(buf):
                 main(args=['https://ttt.com'])
-                self.assertEqual(buf.getvalue(), 'Connection error. Please correct the URL and try again\n')
+                self.assertEqual(buf.getvalue(), 'Invalid URL {url}. Please correct the URL and try again\n')
 
 
 if __name__ == '__main__':
