+++ rss_reader/tests/test_select_from_db.py
@@ -6,24 +6,17 @@ import peewee
 import datetime
 import re
 import json
-from rss_reader.rss_reader import create_db, get_storage, get_arg_parser, get_news_from_storage, print_news
+from rss_reader import reader
 from rss_reader.db_worker import get_path, RssStorage
 from rss_reader.db_worker import db as rss_db
 from contextlib import redirect_stdout
-from bs4 import BeautifulSoup
-
-try:
-    from tests.test_arguments import create_test_data
-
-except ModuleNotFoundError:
-    from test_arguments import create_test_data
 
 
 class TestRssReaderDB(unittest.TestCase):
     @classmethod
     def setUpClass(cls):
-        cls.arg_parser = get_arg_parser()
-        create_db()
+        cls.arg_parser = reader.get_arg_parser()
+        reader.create_db()
         with rss_db.connection_context():
             RssStorage.create_table()
         for item in range(50):
@@ -39,8 +32,6 @@ class TestRssReaderDB(unittest.TestCase):
             except peewee.IntegrityError:
                 pass
 
-        cls.soup = BeautifulSoup(create_test_data(), 'lxml-xml')
-
     @classmethod
     def tearDownClass(cls):
         os.remove(get_path())
@@ -49,11 +40,11 @@ class TestRssReaderDB(unittest.TestCase):
         """Tests --date with limit"""
         args = ['--date=20210530', '--limit=5', ]
         parsed_arg = self.arg_parser.parse_args(args)
-        storage = get_storage(parsed_arg, self.soup)
-        news, items = get_news_from_storage(storage, parsed_arg)
+        storage = reader.selection_from_db(parsed_arg.date)
+        news, items = reader.collect_and_format_news(storage, parsed_arg)
         with io.StringIO() as buf, redirect_stdout(buf):
-            get_storage(parsed_arg)
-            print_news(parsed_arg.json, news)
+            reader.selection_from_db(parsed_arg.date)
+            reader.print_news(parsed_arg.json, news)
             result_news_dates_and_links = re.findall('Date: .+\nLink: .+\n', buf.getvalue())
             self.assertEqual(len(result_news_dates_and_links), 5)
 
@@ -61,11 +52,11 @@ class TestRssReaderDB(unittest.TestCase):
         """Tests --date without limit"""
         args = ['--date=20210530', ]
         parsed_arg = self.arg_parser.parse_args(args)
-        storage = get_storage(parsed_arg, self.soup)
-        news, items = get_news_from_storage(storage, parsed_arg)
+        storage = reader.selection_from_db(parsed_arg.date)
+        news, items = reader.collect_and_format_news(storage, parsed_arg)
         with io.StringIO() as buf, redirect_stdout(buf):
-            get_storage(parsed_arg)
-            print_news(parsed_arg.json, news)
+            reader.selection_from_db(parsed_arg.date)
+            reader.print_news(parsed_arg.json, news)
             result_news_dates_and_links = re.findall('Date: .+\nLink: .+\n', buf.getvalue())
             self.assertEqual(len(result_news_dates_and_links), 50)
 
@@ -73,11 +64,11 @@ class TestRssReaderDB(unittest.TestCase):
         """Tests --date with large limit"""
         args = ['--date=20210530', '--limit=999']
         parsed_arg = self.arg_parser.parse_args(args)
-        storage = get_storage(parsed_arg, self.soup)
-        news, items = get_news_from_storage(storage, parsed_arg)
+        storage = reader.selection_from_db(parsed_arg.date)
+        news, items = reader.collect_and_format_news(storage, parsed_arg)
         with io.StringIO() as buf, redirect_stdout(buf):
-            get_storage(parsed_arg)
-            print_news(parsed_arg.json, news)
+            reader.selection_from_db(parsed_arg.date)
+            reader.print_news(parsed_arg.json, news)
             result_news_dates_and_links = re.findall('Date: .+\nLink: .+\n', buf.getvalue())
             self.assertEqual(len(result_news_dates_and_links), 50)
 
@@ -85,11 +76,11 @@ class TestRssReaderDB(unittest.TestCase):
         """Tests --date with --json printing"""
         args = ['--date=20210530', '--json', ]
         parsed_arg = self.arg_parser.parse_args(args)
-        storage = get_storage(parsed_arg, self.soup)
-        news, items = get_news_from_storage(storage, parsed_arg)
+        storage = reader.selection_from_db(parsed_arg.date)
+        news, items = reader.collect_and_format_news(storage, parsed_arg)
         with io.StringIO() as buf, redirect_stdout(buf):
-            get_storage(parsed_arg)
-            print_news(parsed_arg.json, news)
+            reader.selection_from_db(parsed_arg.date)
+            reader.print_news(parsed_arg.json, news)
             try:
                 json.loads(buf.getvalue())
             except json.JSONDecodeError:
@@ -99,33 +90,25 @@ class TestRssReaderDB(unittest.TestCase):
         """Tests --date with --json without limit"""
         args = ['--date=20210530', '--json', ]
         parsed_arg = self.arg_parser.parse_args(args)
-        storage = get_storage(parsed_arg, self.soup)
-        news, items = get_news_from_storage(storage, parsed_arg)
+        storage = reader.selection_from_db(parsed_arg.date)
+        news, items = reader.collect_and_format_news(storage, parsed_arg)
         with io.StringIO() as buf, redirect_stdout(buf):
-            get_storage(parsed_arg)
-            print_news(parsed_arg.json, news)
-            try:
-                res = json.loads(buf.getvalue())
-            except json.JSONDecodeError:
-                self.fail('JSON Decode Error')
-            else:
-                self.assertEqual(len(res), 50)
+            reader.selection_from_db(parsed_arg.date)
+            reader.print_news(parsed_arg.json, news)
+            res = json.loads(buf.getvalue())
+            self.assertEqual(len(res), 50)
 
     def test_json_news_large_limit(self):
         """Tests --date with --json wit large limit"""
         args = ['--date=20210530', '--json', '--limit=999']
         parsed_arg = self.arg_parser.parse_args(args)
-        storage = get_storage(parsed_arg, self.soup)
-        news, items = get_news_from_storage(storage, parsed_arg)
+        storage = reader.selection_from_db(parsed_arg.date)
+        news, items = reader.collect_and_format_news(storage, parsed_arg)
         with io.StringIO() as buf, redirect_stdout(buf):
-            get_storage(parsed_arg)
-            print_news(parsed_arg.json, news)
-            try:
-                res = json.loads(buf.getvalue())
-            except json.JSONDecodeError:
-                self.fail('JSON Decode Error')
-            else:
-                self.assertEqual(len(res), 50)
+            reader.selection_from_db(parsed_arg.date)
+            reader.print_news(parsed_arg.json, news)
+            res = json.loads(buf.getvalue())
+            self.assertEqual(len(res), 50)
 
 
 if __name__ == '__main__':
